Use `Long` consistently to define body parser buffer length (#8627)
## Purpose

This is a follow up to PR #8520 to use `Long` consistently to define body parser buffer length.

## References

See conversation starting here: https://github.com/playframework/playframework/pull/8520#issuecomment-405956897

+++ framework/src/play/src/main/scala/play/api/http/HttpConfiguration.scala
@@ -128,7 +128,7 @@ case class FlashConfiguration(
  * @param maxDiskBuffer   The maximum size that a request body should be buffered on disk.
  */
 case class ParserConfiguration(
-    maxMemoryBuffer: Int = 102400,
+    maxMemoryBuffer: Long = 102400,
     maxDiskBuffer: Long = 10485760)
 
 /**
@@ -190,8 +190,7 @@ object HttpConfiguration {
     HttpConfiguration(
       context = context,
       parser = ParserConfiguration(
-        maxMemoryBuffer = config.getDeprecated[ConfigMemorySize]("play.http.parser.maxMemoryBuffer", "parsers.text.maxLength")
-          .toBytes.toInt,
+        maxMemoryBuffer = config.getDeprecated[ConfigMemorySize]("play.http.parser.maxMemoryBuffer", "parsers.text.maxLength").toBytes,
         maxDiskBuffer = config.get[ConfigMemorySize]("play.http.parser.maxDiskBuffer").toBytes
       ),
       actionComposition = ActionCompositionConfiguration(
